[INFRA] split minimiser model
