+++ src/search/compute_simple_model.cpp
@@ -1,0 +1,24 @@
+#include <raptor/search/compute_simple_model.hpp>
+
+namespace raptor
+{
+
+std::vector<size_t> compute_simple_model(search_arguments const & arguments)
+{
+    std::vector<size_t> precomp_thresholds;
+
+    if (!arguments.threshold && !do_cerealisation_in(precomp_thresholds, arguments))
+    {
+        precomp_thresholds = precompute_threshold(arguments.pattern_size,
+                                                  arguments.window_size,
+                                                  arguments.kmer_size,
+                                                  arguments.errors,
+                                                  arguments.tau);
+
+        do_cerealisation_out(precomp_thresholds, arguments);
+    }
+
+    return precomp_thresholds;
+}
+
+} // namespace raptor
